requirements.txt
requests
python-dotenv
beautifulsoup4
pandas
pytest
PyYAML
uvicorn
mcp-redmine==2025.04.09.153531
adk-python
bs4
